game-samples/requirements.txt
google-cloud-spanner==3.14.0
locust==2.8.6
